+++ apis/r/R/RcppExports.R
@@ -9,8 +9,8 @@ createSchemaFromArrow <- function(uri, nasp, nadimap, nadimsp, sparse, datatype,
     invisible(.Call(`_tiledbsoma_createSchemaFromArrow`, uri, nasp, nadimap, nadimsp, sparse, datatype, pclst, ctxxp, tsvec))
 }
 
-writeArrayFromArrow <- function(uri, naap, nasp, arraytype = "", config = NULL, tsvec = NULL) {
-    invisible(.Call(`_tiledbsoma_writeArrayFromArrow`, uri, naap, nasp, arraytype, config, tsvec))
+writeArrayFromArrow <- function(uri, naap, nasp, ctxxp, arraytype = "", config = NULL, tsvec = NULL) {
+    invisible(.Call(`_tiledbsoma_writeArrayFromArrow`, uri, naap, nasp, ctxxp, arraytype, config, tsvec))
 }
 
 #' @noRd
@@ -60,6 +60,7 @@ c_group_put_metadata <- function(xp, key, obj) {
 
 #' Get nnumber of metadata items
 #' @param uri The array URI
+#' @param is_array A boolean to indicate array or group
 #' @param ctxxp An external pointer to the SOMAContext wrapper
 #' @export
 get_metadata_num <- function(uri, is_array, ctxxp) {
@@ -71,6 +72,7 @@ get_metadata_num <- function(uri, is_array, ctxxp) {
 #' This function currently supports metadata as either a string or an 'int64' (or 'int32').
 #' It will error if a different datatype is encountered.
 #' @param uri The array URI
+#' @param is_array A boolean to indicate array or group
 #' @param ctxxp An external pointer to the SOMAContext wrapper
 #' @export
 get_all_metadata <- function(uri, is_array, ctxxp) {
@@ -81,6 +83,7 @@ get_all_metadata <- function(uri, is_array, ctxxp) {
 #'
 #' @param uri The array URI
 #' @param key The array metadata key
+#' @param is_array A boolean to indicate array or group
 #' @param ctxxp An external pointer to the SOMAContext wrapper
 #' @export
 get_metadata <- function(uri, key, is_array, ctxxp) {
@@ -91,6 +94,7 @@ get_metadata <- function(uri, key, is_array, ctxxp) {
 #'
 #' @param uri The array URI
 #' @param key The array metadata key
+#' @param is_array A boolean to indicate array or group
 #' @param ctxxp An external pointer to the SOMAContext wrapper
 #' @export
 has_metadata <- function(uri, key, is_array, ctxxp) {
@@ -101,6 +105,7 @@ has_metadata <- function(uri, key, is_array, ctxxp) {
 #'
 #' @param uri The array URI
 #' @param key The array metadata key
+#' @param is_array A boolean to indicate array or group
 #' @param ctxxp An external pointer to the SOMAContext wrapper
 #' @export
 delete_metadata <- function(uri, key, is_array, ctxxp) {
@@ -134,8 +139,8 @@ reindex_lookup <- function(idx, kvec) {
 }
 
 #' @noRd
-soma_array_reader_impl <- function(uri, colnames = NULL, qc = NULL, dim_points = NULL, dim_ranges = NULL, batch_size = "auto", result_order = "auto", loglevel = "auto", config = NULL, timestamprange = NULL) {
-    .Call(`_tiledbsoma_soma_array_reader`, uri, colnames, qc, dim_points, dim_ranges, batch_size, result_order, loglevel, config, timestamprange)
+soma_array_reader_impl <- function(uri, ctxxp, colnames = NULL, qc = NULL, dim_points = NULL, dim_ranges = NULL, batch_size = "auto", result_order = "auto", loglevel = "auto", timestamprange = NULL) {
+    .Call(`_tiledbsoma_soma_array_reader`, uri, ctxxp, colnames, qc, dim_points, dim_ranges, batch_size, result_order, loglevel, timestamprange)
 }
 
 #' Set the logging level for the R package and underlying C++ library
@@ -153,8 +158,8 @@ get_column_types <- function(uri, colnames) {
     .Call(`_tiledbsoma_get_column_types`, uri, colnames)
 }
 
-nnz <- function(uri, config = NULL) {
-    .Call(`_tiledbsoma_nnz`, uri, config)
+nnz <- function(uri, ctxxp) {
+    .Call(`_tiledbsoma_nnz`, uri, ctxxp)
 }
 
 #' @noRd
@@ -167,24 +172,32 @@ check_arrow_array_tag <- function(xp) {
     .Call(`_tiledbsoma_check_arrow_array_tag`, xp)
 }
 
-shape <- function(uri, config = NULL) {
-    .Call(`_tiledbsoma_shape`, uri, config)
+shape <- function(uri, ctxxp) {
+    .Call(`_tiledbsoma_shape`, uri, ctxxp)
 }
 
-maxshape <- function(uri, config = NULL) {
-    .Call(`_tiledbsoma_maxshape`, uri, config)
+maxshape <- function(uri, ctxxp) {
+    .Call(`_tiledbsoma_maxshape`, uri, ctxxp)
 }
 
-maybe_soma_joinid_shape <- function(uri, config = NULL) {
-    .Call(`_tiledbsoma_maybe_soma_joinid_shape`, uri, config)
+maybe_soma_joinid_shape <- function(uri, ctxxp) {
+    .Call(`_tiledbsoma_maybe_soma_joinid_shape`, uri, ctxxp)
 }
 
-maybe_soma_joinid_maxshape <- function(uri, config = NULL) {
-    .Call(`_tiledbsoma_maybe_soma_joinid_maxshape`, uri, config)
+maybe_soma_joinid_maxshape <- function(uri, ctxxp) {
+    .Call(`_tiledbsoma_maybe_soma_joinid_maxshape`, uri, ctxxp)
 }
 
-has_current_domain <- function(uri, config = NULL) {
-    .Call(`_tiledbsoma_has_current_domain`, uri, config)
+has_current_domain <- function(uri, ctxxp) {
+    .Call(`_tiledbsoma_has_current_domain`, uri, ctxxp)
+}
+
+resize <- function(uri, new_shape, ctxxp) {
+    invisible(.Call(`_tiledbsoma_resize`, uri, new_shape, ctxxp))
+}
+
+tiledbsoma_upgrade_shape <- function(uri, new_shape, ctxxp) {
+    invisible(.Call(`_tiledbsoma_tiledbsoma_upgrade_shape`, uri, new_shape, ctxxp))
 }
 
 #' Iterator-Style Access to SOMA Array via SOMAArray
@@ -221,8 +234,8 @@ has_current_domain <- function(uri, config = NULL) {
 #' @examples
 #' \dontrun{
 #' uri <- extract_dataset("soma-dataframe-pbmc3k-processed-obs")
-#' ctx <- tiledb::tiledb_ctx()
-#' sr <- sr_setup(uri, config=as.character(tiledb::config(ctx)))
+#' ctxcp <- soma_context()
+#' sr <- sr_setup(uri, ctxxp)
 #' rl <- data.frame()
 #' while (!sr_complete(sr)) {
 #'   dat <- sr_next(sr)
@@ -232,8 +245,8 @@ has_current_domain <- function(uri, config = NULL) {
 #' summary(rl)
 #' }
 #' @noRd
-sr_setup <- function(uri, config, colnames = NULL, qc = NULL, dim_points = NULL, dim_ranges = NULL, batch_size = "auto", result_order = "auto", timestamprange = NULL, loglevel = "auto") {
-    .Call(`_tiledbsoma_sr_setup`, uri, config, colnames, qc, dim_points, dim_ranges, batch_size, result_order, timestamprange, loglevel)
+sr_setup <- function(uri, ctxxp, colnames = NULL, qc = NULL, dim_points = NULL, dim_ranges = NULL, batch_size = "auto", result_order = "auto", timestamprange = NULL, loglevel = "auto") {
+    .Call(`_tiledbsoma_sr_setup`, uri, ctxxp, colnames, qc, dim_points, dim_ranges, batch_size, result_order, timestamprange, loglevel)
 }
 
 sr_complete <- function(sr) {