+++ apis/r/R/TileDBArray.R
@@ -154,10 +154,7 @@ TileDBArray <- R6::R6Class(
     #' (lifecycle: maturing)
     #' @return A named vector of dimension length (and the same type as the dimension)
     shape = function() {
-      as.integer64(shape(
-        self$uri,
-        config=as.character(tiledb::config(self$tiledbsoma_ctx$context()))
-      ))
+      as.integer64(shape(self$uri, private$.soma_context))
     },
 
     #' @description Retrieve the maxshape, i.e. maximum possible that the
@@ -165,10 +162,7 @@ TileDBArray <- R6::R6Class(
     #' (lifecycle: maturing)
     #' @return A named vector of dimension length (and the same type as the dimension)
     maxshape = function() {
-      as.integer64(maxshape(
-        self$uri,
-        config=as.character(tiledb::config(self$tiledbsoma_ctx$context()))
-      ))
+      as.integer64(maxshape(self$uri, private$.soma_context))
     },
 
     #' @description Retrieve the range of indexes for a dimension that were