+++ apis/r/R/TileDBObject.R
@@ -15,7 +15,8 @@ TileDBObject <- R6::R6Class(
     #' @param internal_use_only Character value to signal this is a 'permitted' call,
     #' as `new()` is considered internal and should not be called directly.
     initialize = function(uri, platform_config = NULL, tiledbsoma_ctx = NULL,
-                          tiledb_timestamp = NULL, internal_use_only = NULL) {
+                          tiledb_timestamp = NULL, internal_use_only = NULL,
+                          soma_context = NULL) {
       if (is.null(internal_use_only) || internal_use_only != "allowed_use") {
         stop(paste("Use of the new() method is for internal use only. Consider using a",
                    "factory method as e.g. 'SOMADataFrameOpen()'."), call. = FALSE)
@@ -37,6 +38,17 @@ TileDBObject <- R6::R6Class(
       }
       private$.tiledbsoma_ctx <- tiledbsoma_ctx
       private$.tiledb_ctx <- self$tiledbsoma_ctx$context()
+
+      # TODO: re-enable once new UX is worked out
+      # soma_context <- soma_context %||% soma_context()
+      # stopifnot(
+      #   "'soma_context' must be a pointer" = inherits(x = soma_context, what = 'externalptr')
+      # )
+      if (is.null(soma_context)) {
+          private$.soma_context <- soma_context()  # FIXME via factory and paramater_config
+      } else {
+          private$.soma_context <- soma_context
+      }
 
       if (!is.null(tiledb_timestamp)) {
         stopifnot(
@@ -185,6 +197,9 @@ TileDBObject <- R6::R6Class(
     # * In particular, an is-open predicate can be reliably implemented by
     #   checking if .mode is non-null.
     .mode = NULL,
+
+    ## soma_context
+    .soma_context = NULL,
 
     # @description Contains TileDBURI object
     tiledb_uri = NULL,