+++ apis/r/NEWS.md
@@ -2,10 +2,13 @@
 
 ## Changes
 
+* New `resize` and `tiledbsoma_upgrade_shape` accessors as part of the [new-shape project](https://github.com/single-cell-data/TileDB-SOMA/issues/2407)
 * Make use of timestamp ranges in libtiledbsoma
 * Simplify timestamp ranges; strengthen assumptions about `tiledb_timestamp`
 * Use cached timestamps in `$write()` and `$create()`
 * Fix bug in blockwise iteration
+* Lay groundwork for cached SOMA contexts within objects rather than re-creating contexts
+* SOMA context objects are used throughout SOMA object creation
 
 # tiledbsoma 1.13.0